Smart_Testcase_creator_Jira_Integration_agent/requirements.txt
streamlit
requests
groq
python-dotenv
keyring
